+++ src/imio/smartweb/common/adapters.py
@@ -1,7 +1,13 @@
 # -*- coding: utf-8 -*-
 
+from imio.smartweb.common.utils import get_image_format
 from plone.namedfile.interfaces import IAvailableSizes
+from plone.namedfile.interfaces import INamedImageField
+from plone.namedfile.field import InvalidImageFile
+from plone.namedfile.utils import get_contenttype
+from zope.component import adapter
 from zope.component import getUtility
+from zope.interface import Interface
 
 
 class BaseCroppingProvider(object):
@@ -17,3 +23,25 @@ class BaseCroppingProvider(object):
         scales = list(allowed_sizes.keys())
         scales.remove("banner")
         return scales
+
+
+@adapter(INamedImageField, Interface)
+class ImageContenttypeValidator:
+    def __init__(self, field, value):
+        self.field = field
+        self.value = value
+
+    def __call__(self):
+        if self.value is None:
+            return
+        mimetype = get_image_format(self.value)
+
+        valid_mimetypes = [
+            "image/gif",
+            "image/jpeg",
+            "image/png",
+            "image/svg+xml",
+            "image/webp",
+        ]
+        if mimetype not in valid_mimetypes:
+            raise InvalidImageFile(mimetype, self.field.__name__)