+++ src/imio/smartweb/common/browser/cropping.py
@@ -51,8 +51,7 @@ class SmartwebCroppingImageScalingFactory(CroppingImageScalingFactory):
     ):
         storage = Storage(self.context)
         self.box = storage.read(fieldname, scale)
-
-        if "portrait" in scale or "paysage" in scale or "carre" in scale:
+        if scale and ("portrait" in scale or "paysage" in scale or "carre" in scale):
             orientation = scale.split("_")[0]
             # take cropping box from "affiche" scale
             self.box = storage.read(fieldname, f"{orientation}_affiche")